requirements.txt
tqdm==4.19.2
torchvision==0.2.1
numpy
torch==0.3.1
matplotlib==2.2.2
Pillow==5.1.0
scikit_learn==0.19.1
vtk